requirements.txt
fastapi
uvicorn
pandas
numpy
nltk
sentence-transformers
transformers
torch
bitsandbytes
langchain
langchain-community
faiss-cpu
huggingface-hub
streamlit
accelerate
requests
python-dotenv
pydantic
openai
